Gestion des erreurs sur les formulaires d'ajout / update. Touch-ups au design pour essayer d'uniformiser un peu (reste quelques trucs à faire). Màj pour les images de catégories : elles sont maintenant stockées directement dans la bdd. Un bouton dans le coin droit de la carte (accessible lors du clic sur 'modifier') permet de changer dynamiquement l'image associée à une catégorie. Idem on peut désormais initialiser une catégorie avec une image grâce à un bouton similaire. Mise en place d'un attribut 'taken' dans les commandes qui représente simplement si une commande est en train d'être réalisée (à la demande de l'équipe drinkiit). Cet attribut est représenté sous la forme d'une check box. Dès qu'elle est cochée, un appel ajax est fait automatiquement pour mettre à jour la commande (pour pas que la checkbox se décoche si on refresh la page). De même, la page des commandes se rafrâichit désormais automatiquement toutes les deux minutes (interval final à déterminer avec Nicolas). Le site possède désormais toutes les fonctionnalités du site précédent ! (minus l'API)

+++ app/controllers/orders_controller.rb
@@ -1,5 +1,5 @@
 class OrdersController < ApplicationController
-before_action :set_order, only: [:validate, :destroy]
+before_action :set_order, only: [:take, :validate, :destroy_in_orders, :destroy_in_my_orders]
 
 	def index
 		if signed_in? && current_user.admin
@@ -7,6 +7,10 @@ before_action :set_order, only: [:validate, :destroy]
 		else
 			redirect_to root_path
 		end
+	end
+
+	def refresh
+		set_order_list
 	end
 
 	def show
@@ -18,27 +22,42 @@ before_action :set_order, only: [:validate, :destroy]
 	end
 
 	def create
-		@success = true
-		@message = "Commande effectuée avec succès !"
+		set_success(true, "Commande effectuée avec succès !")
 		if signed_in?
 		    @order = Order.new(order_params)		    
 		    @meal = Meal.find(@order.meal_id)
+		    if @order.quantity.nil?
+		    	set_success(false, "La quantité doit être rempli(e)")
+		    	return
+		    elsif @order.quantity < 1
+		    	set_success(false, "La quantité doit être au moins égale à 1 !")
+		    	return
+		    end
 		    @order.total = @order.quantity * @meal.price
+		    @order.taken = false
+
+		    if !params[:ingredients].nil?
+		    	params[:ingredients].each do |type_id, ingredient_id|
+		    		if ingredient_id.empty? || ingredient_id.nil?
+		    			set_success(false, IngredientType.find(type_id).name + " non renseigné(e) !")
+		    			return
+		    		else
+		    			@order.ingredients << Ingredient.find(ingredient_id)
+		    		end
+		    	end
+		    end
 		    if current_user.solde - @order.total >= 0
 		    	if check_user_for(@order)			    
 				    @order.served = false
 				    @order.ordered_at = DateTime.now
 				    if !@order.save
-				    	@success = false
-				    	@message = "Une erreur est survenue"
+				    	set_success(false, get_error_message(@order.errors, "Order"))
 			    	end
 				else
-					@success = false
-					@message = "Erreur d'authentification !"
+					set_success(false, "Erreur d'authentification !")
 				end
 			else
-				@success = false
-				@message = "Vous n'avez pas assez d'argent !"
+				set_success(false, "Vous n'avez pas assez d'argent !")
 		    end 
 		else
 			redirect_to root_path
@@ -46,25 +65,42 @@ before_action :set_order, only: [:validate, :destroy]
   	end
 
   	def validate
-  		@success = true
-  		@message = "Commande archivée !"
+  		set_success(true, "Commande archivée !")
   		@user = @order.user
   		new_solde = @user.solde - @order.total
   		if !(@user.update_attribute('solde', new_solde) && @order.update_attribute('served', true))
-  			@success = false
-  			@message = "Une erreur est survenue"
+  			if @user.errors.any?
+  				set_success(false, get_error_message(@user.errors, "User"))
+  			else
+  				set_success(false, get_error_message(@order.errors, "Order"))
+  			end
   		else
   			set_order_list
 		end
   	end
 
-  	def destroy
-  		if signed_in?
-	  		@success = true
-	  		@message = "Commande supprimée !"
+  	def take
+  		@order.update_attribute('taken', order_params[:taken])
+  	end
+
+  	def destroy_in_orders
+  		if signed_in? && current_user.admin
+	  		set_success(true, "Commande supprimée !")
 	  		if !@order.destroy
-	  			@success = false
-	  			@message = "Une erreur s'est produite lors de la suppression"
+	  			set_success(false, get_error_message(@order.errors, "Order"))
+	  		else
+	  			set_order_list
+	  		end
+	  	else
+	  		redirect_to root_path
+	  	end
+  	end
+
+  	def destroy_in_my_orders
+  		if signed_in?
+	  		set_success(true, "Commande supprimée !")
+	  		if !@order.destroy
+	  			set_success(false, get_error_message(@order.errors, "Order"))
 	  		else
 	  			set_my_orders
 	  		end
@@ -72,7 +108,6 @@ before_action :set_order, only: [:validate, :destroy]
 	  		redirect_to root_path
 	  	end
   	end
-
 	private
 
 	def set_order
@@ -80,7 +115,7 @@ before_action :set_order, only: [:validate, :destroy]
 	end
 
 	def set_order_list
-		@orders = Order.order('ordered_at')
+		@orders = Order.order('ordered_at DESC')
 		@today = @orders.select { |order| order.served == false && order.ordered_at.to_date == DateTime.now.to_date }
 		@historique = @orders.select { |order| order.served == true || order.ordered_at.to_date != DateTime.now.to_date }
 	end
@@ -95,8 +130,7 @@ before_action :set_order, only: [:validate, :destroy]
 	end
 
 	def order_params
-      allow = [:quantity, :meal_id, :user_id, :comment]
+      allow = [:quantity, :meal_id, :user_id, :comment, :taken]
       params.require(:order).permit(allow)
     end
-
 end
